Added sanity check
If you input the wrong module, the system will let it pass but it will not be visible - the result is that you will have claims for the wrong module you cannot delete. And you cannot claim for the correct module because the wrong module will put too many hours in the system.